requirements.txt
streamlit
langchain~=0.3.25
langchain-litellm==0.2.0
langchain-community
python-dotenv
PyYAML
requests